refactor: Update Navbar component to include logo image in mobile view
The Navbar component was modified to include the logo image in the mobile view. The previous code that was commented out has been replaced with a new code block that displays the logo using the Image component. This change improves the visual appearance and branding of the application in the mobile view.